COM-1522: rename files
