+++ src/components/cards/ProgressBar/style.ts
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
 import { FIRA_SANS_BOLD } from '../../../styles/fonts';
 import { GREY, YELLOW } from '../../../styles/colors';
-import { BORDER_RADIUS, BORDER_WIDTH, MARGIN } from '../../../styles/metrics';
+import { BORDER_RADIUS, BORDER_WIDTH, MARGIN, PROGRESS_BAR_HEIGHT } from '../../../styles/metrics';
 
 const styles = (progressPercentage: number) => StyleSheet.create({
   container: {
@@ -10,7 +10,7 @@ const styles = (progressPercentage: number) => StyleSheet.create({
     flexDirection: 'row',
     backgroundColor: GREY[200],
     borderRadius: BORDER_RADIUS.SM,
-    height: 8,
+    height: PROGRESS_BAR_HEIGHT,
     flex: 1,
     justifyContent: 'flex-start',
   },
